refactor(impersonate): Reuse code
