+++ src/impersonate/profile/chrome/v106.rs
@@ -1,8 +1,5 @@
 use std::sync::Arc;
 
-use boring::ssl::{
-    CertCompressionAlgorithm, SslConnector, SslConnectorBuilder, SslMethod, SslVersion,
-};
 use http::{
     header::{
         ACCEPT, ACCEPT_ENCODING, ACCEPT_LANGUAGE, DNT, UPGRADE_INSECURE_REQUESTS, USER_AGENT,
@@ -11,6 +8,8 @@ use http::{
 };
 
 use crate::impersonate::{Http2Data, ImpersonateSettings};
+
+use super::create_ssl_connector;
 
 pub(crate) fn get_settings(headers: HeaderMap) -> ImpersonateSettings {
     ImpersonateSettings {
@@ -27,71 +26,6 @@ pub(crate) fn get_settings(headers: HeaderMap) -> ImpersonateSettings {
         gzip: true,
         brotli: true,
     }
-}
-
-fn create_ssl_connector(h2: bool) -> SslConnectorBuilder {
-    let mut builder = SslConnector::builder(SslMethod::tls_client()).unwrap();
-
-    builder.set_default_verify_paths().unwrap();
-
-    builder.set_grease_enabled(true);
-
-    builder.enable_ocsp_stapling();
-
-    let cipher_list = [
-        "TLS_AES_128_GCM_SHA256",
-        "TLS_AES_256_GCM_SHA384",
-        "TLS_CHACHA20_POLY1305_SHA256",
-        "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256",
-        "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
-        "TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384",
-        "TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384",
-        "TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256",
-        "TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256",
-        "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA",
-        "TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA",
-        "TLS_RSA_WITH_AES_128_GCM_SHA256",
-        "TLS_RSA_WITH_AES_256_GCM_SHA384",
-        "TLS_RSA_WITH_AES_128_CBC_SHA",
-        "TLS_RSA_WITH_AES_256_CBC_SHA",
-    ];
-
-    builder.set_cipher_list(&cipher_list.join(":")).unwrap();
-
-    let sigalgs_list = [
-        "ecdsa_secp256r1_sha256",
-        "rsa_pss_rsae_sha256",
-        "rsa_pkcs1_sha256",
-        "ecdsa_secp384r1_sha384",
-        "rsa_pss_rsae_sha384",
-        "rsa_pkcs1_sha384",
-        "rsa_pss_rsae_sha512",
-        "rsa_pkcs1_sha512",
-    ];
-
-    builder.set_sigalgs_list(&sigalgs_list.join(":")).unwrap();
-
-    builder.enable_signed_cert_timestamps();
-
-    if h2 {
-        builder.set_alpn_protos(b"\x02h2\x08http/1.1").unwrap();
-    } else {
-        builder.set_alpn_protos(b"\x08http/1.1").unwrap();
-    }
-
-    builder
-        .add_cert_compression_alg(CertCompressionAlgorithm::Brotli)
-        .unwrap();
-
-    builder
-        .set_min_proto_version(Some(SslVersion::TLS1_2))
-        .unwrap();
-
-    builder
-        .set_max_proto_version(Some(SslVersion::TLS1_3))
-        .unwrap();
-
-    builder
 }
 
 fn create_headers(mut headers: HeaderMap) -> HeaderMap {
